refactor common utilities for the pylibraft tests
